部门管理页面功能

+++ studying-web/src/main/java/com/small/studyingweb/controller/system/SysDepartmentController.java
@@ -24,10 +24,10 @@ public class SysDepartmentController extends BaseController {
 
     @GetMapping("/dynamicGetDept")
     @ResponseBody
-    public List<SysDepartment> dynamicGetDeptByPId(@RequestParam(value = "departmentId",required = false)Integer pId){
+    public List<SysDepartment> dynamicGetDeptByPId(@RequestParam(value = "departmentId",required = false)Long pId){
         SysDepartmentQuery query = new SysDepartmentQuery();
         if(pId==null){
-            query.setParentId(0);
+            query.setParentId(0L);
         }else{
             query.setParentId(pId);
         }
